+++ R/server.R
@@ -52,7 +52,7 @@ server <- function(input, output) {
                         which(MCE$test$item_options[[pick]] %in% ip) - 1L
                     if(!is.na(MCE$test$item_answers[[pick]]) && 
                            MCE$test$item_class[pick] != 'nestlogit')
-                        MCE$person$responses[pick] <- as.integer(ip == MCE$test$item_answers[[pick]])
+                        MCE$person$responses[pick] <- as.integer(ip %in% MCE$test$item_answers[[pick]])
                     
                     MCE$person$item_time[pick] <- proc.time()[3L] - MCE$start_time - 
                         sum(MCE$person$item_time)